Deleted wrong solution. Were corrected files
