--- Lesson8/src/main/java/realisationN1/service/Post.java
@@ -1,15 +1,0 @@
-package realisationN1.service;
-
-public enum Post {
-    DIRECTOR(2.5f),
-    WORKER(1.3f);
-    private float coef;
-
-    Post(float coef) {
-        this.coef = coef;
-    }
-
-    public float getCoef() {
-        return coef;
-    }
-}
